requirements.txt
langchain
langgraph
langchain-community
langchain-openai
langchain-groq
langchain-google-genai
langchain-chroma
chromadb
openai
tiktoken
sentence-transformers
fastapi
gradio
streamlit
uvicorn
python-dotenv
pydantic
pydantic-settings
requests
beautifulsoup4
lxml
PyMuPDF
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
httpx
aiohttp
rich
pytest
pytest-asyncio
deep-translator==1.11.4
validators